A python minigame for Android dimensions
Developed on Pydroid (Python for android), player guesses a number and a random number for the opponent is also generated. Both numbers are compared against the "winning" number that is also randomly generated, and the number that is the closest wins, and the line is drawn to represent the score.